update cookie service, and redirect to login for add to cart

+++ src/app/components/pages/wishlist/wishlist.component.ts
@@ -5,6 +5,10 @@ import { Product } from './../../../modals/product.model';
 import { CartService } from '../../shared/services/cart.service';
 import { WishlistService } from '../../shared/services/wishlist.service';
 import { environment } from 'src/environments/environment';
+import { ConfirmationDialogFrontComponent } from '../../shared/confirmation-dialog-front/confirmation-dialog-front.component';
+import { LoggedInUserService } from 'src/app/core/services/loggedInUser/logged-in-user.service';
+import { Router } from '@angular/router';
+import { MatDialog } from '@angular/material/dialog';
 
 @Component({
   selector: 'app-wishlist',
@@ -17,7 +21,9 @@ export class WishlistComponent implements OnInit {
 
   constructor(
     private cartService: CartService,
+    private loggedInUserService: LoggedInUserService,
     private wishlistService: WishlistService,
+    private router: Router,
     private metaService: MetaService,
   ) {
     this.product = this.wishlistService.getProducts();
@@ -32,15 +38,18 @@ export class WishlistComponent implements OnInit {
     );
   }
 
-  ngOnInit() {
-  }
+  ngOnInit() {}
 
   // Add to cart
   public addToCart(product: Product, quantity: number = 1) {
-    if (quantity > 0) {
-      this.cartService.addToCart(product, quantity);
+    if (this.loggedInUserService.getLoggedInUser()) {
+      if (quantity > 0) {
+        this.cartService.addToCart(product, quantity);
+      }
+      this.wishlistService.removeFromWishlist(product);
+    } else {
+      this.cartService.redirectToLoginWithOrigin(this.router.routerState.snapshot.url);
     }
-    this.wishlistService.removeFromWishlist(product);
   }
 
   // Remove from wishlist
